back/requirements.txt
Flask==3.*
Flask-JWT-Extended
Flask-Cors
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
requests
gunicorn
psycopg2-binary
google-generativeai